Refactor Docker setup and development instructions
This commit refactors the README.md file to update the Docker setup section. It also improves the development setup and installation instructions by using pnpm instead of npm for installing dependencies. Additionally, it adds an option to use localstorage instead of SQLLite for persisting data in the browser. The commit also includes a step to push the database schema when using SQLLite. The contribution guidelines have been removed from the README.md file. Lastly, the commit removes a demo GIF that was previously included.